requirements.txt
# local package
-e .

# external requirements
click
Sphinx
awscli
python-dotenv>=0.5.1
requests
streamlit
